+++ packages/orchestrator/src/config.ts
@@ -15,6 +15,7 @@
 */
 import composerUrl from '@micro-lc/composer/composer.production?url'
 import type {
+  Application,
   Config,
   Content,
   GlobalImportMap,
@@ -25,6 +26,7 @@ import type {
 import page401Url from './assets/401.html?url'
 import page404Url from './assets/404.html?url'
 import page500Url from './assets/500.html?url'
+import { computeAbsoluteRoute } from './web-component/lib'
 
 export const MICRO_LC_MOUNT_POINT = '__MICRO_LC_MOUNT_POINT'
 
@@ -65,11 +67,38 @@ export const defaultConfig: CompleteConfig = {
   version: 2,
 }
 
-export function mergeConfig(input: Config): CompleteConfig {
+const patchApplicationsOntoDocument = (document: Document, applications: Record<string, Application>) =>
+  Object
+    .entries(applications)
+    .reduce<Record<string, Application>>((apps, [key, nextApp]) => {
+      if ('route' in nextApp) {
+        apps[key] = {
+          ...nextApp,
+          route: computeAbsoluteRoute(document, nextApp.route),
+        }
+      }
+
+      if ('config' in nextApp && typeof nextApp.config === 'string') {
+        apps[key] = {
+          ...nextApp,
+          config: computeAbsoluteRoute(document, nextApp.config),
+        }
+      }
+
+      return apps
+    }, applications)
+
+export function mergeConfig(document: Document, input: Config): CompleteConfig {
   const def = defaultConfig
+  const defaultUrl = computeAbsoluteRoute(
+    document, input.settings?.defaultUrl ?? def.settings.defaultUrl
+  )
+  const applications = patchApplicationsOntoDocument(
+    document, input.applications ?? def.applications
+  )
 
   return {
-    applications: input.applications ?? def.applications,
+    applications,
     importmap: input.importmap ?? def.importmap,
     layout: {
       ...input.layout,
@@ -79,7 +108,7 @@ export function mergeConfig(input: Config): CompleteConfig {
       '4xx': input.settings?.['4xx'] ?? def.settings['4xx'],
       '5xx': input.settings?.['5xx'] ?? def.settings['5xx'],
       composerUri: input.settings?.composerUri ?? def.settings.composerUri,
-      defaultUrl: input.settings?.defaultUrl ?? def.settings.defaultUrl,
+      defaultUrl,
       mountPoint: input.settings?.mountPoint ?? def.settings.mountPoint,
       mountPointSelector: input.settings?.mountPointSelector ?? def.settings.mountPointSelector,
     },